refactor: move internal utils and add utils to bundle entries
