+++ src/commands/analyze.ts
@@ -5,7 +5,7 @@ import { listen } from 'listhen'
 import type { NuxtAnalyzeMeta } from '@nuxt/schema'
 import { defu } from 'defu'
 import { loadKit } from '../utils/kit'
-import { clearDir } from '../utils/fs'
+import { clearDir } from '../utils/internal/fs'
 import { overrideEnv } from '../utils/env'
 import { defineCommand } from 'citty'
 import { sharedArgs, legacyRootDirArgs } from './_shared'
