+++ src/commands/build.ts
@@ -2,7 +2,7 @@ import { relative, resolve } from 'pathe'
 import { consola } from 'consola'
 import type { Nitro } from 'nitropack'
 import { loadKit } from '../utils/kit'
-import { clearBuildDir } from '../utils/fs'
+import { clearBuildDir } from '../utils/internal/fs'
 import { overrideEnv } from '../utils/env'
 import { showVersions } from '../utils/banner'
 import { defineCommand } from 'citty'
