+++ src/commands/prepare.ts
@@ -3,7 +3,7 @@ import { consola } from 'consola'
 // we are deliberately inlining this code as a backup in case user has `@nuxt/schema<3.7`
 import { writeTypes as writeTypesLegacy } from '@nuxt/kit'
 
-import { clearBuildDir } from '../utils/fs'
+import { clearBuildDir } from '../utils/internal/fs'
 import { loadKit } from '../utils/kit'
 import { defineCommand } from 'citty'
