+++ src/utils/index.ts
@@ -1,13 +1,10 @@
 export * from './banner'
-export * from './cjs'
 export * from './console'
 export * from './dev'
 export * from './engines'
 export * from './env'
 export * from './esm'
-export * from './fs'
 export * from './kit'
 export * from './nuxt'
-export * from './packageManagers'
 export * from './templates'
 export * from './update'
